Squashed commit of the following:

commit b70a40dbf2ad4915599a52cac9f2e9c6012e8f2e
Author: Yuduo Wu <yuduo@apple.com>
Date:   Fri Jun 12 13:09:38 2020 -0700

    Clean some more warnings

+++ coremltools/converters/mil/mil/ops/defs/tensor_operation.py
@@ -324,7 +324,7 @@ class range_1d(Operation):
         start = self.start.val
         end   = self.end.val
         step  = self.step.val
-        return np.arange(start, end, step)
+        return np.arange(start, end, step).astype(np.int32)
 
     def type_inference(self):
         start = self.start.sym_val
